Creamos todas las entidades sin implementar. Implementamos el matchmacking (CRUD, ENDPOINTS)

+++ donaciones/src/test/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/SegmentacionDonacionTest.java
@@ -5,6 +5,14 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
+import ar.edu.utn.frba.ddsi.donaciones.models.entities.donaciones.Bien;
+import ar.edu.utn.frba.ddsi.donaciones.models.entities.donaciones.Categoria;
+import ar.edu.utn.frba.ddsi.donaciones.models.entities.donaciones.Donacion;
+import ar.edu.utn.frba.ddsi.donaciones.models.entities.donaciones.EstadoBien;
+import ar.edu.utn.frba.ddsi.donaciones.models.entities.donaciones.RegistroDonacion;
+import ar.edu.utn.frba.ddsi.donaciones.models.entities.donaciones.Subcategoria;
+import ar.edu.utn.frba.ddsi.donaciones.models.entities.donaciones.TipoEstadoDonacion;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -15,7 +23,8 @@ import static org.junit.jupiter.api.Assertions.*;
 /**
  * Tests para el Requerimiento: "Registrar donación de bienes"
  * Los bienes de una donación deben segmentarse automáticamente en Donaciones
- * agrupadas por subcategoría, estado del bien y fecha de vencimiento (si aplica).
+ * agrupadas por subcategoría, estado del bien y fecha de vencimiento (si
+ * aplica).
  * Cada Donacion resultante arranca en estado EN_DEPOSITO.
  */
 @DisplayName("Segmentación automática de donaciones (RegistroDonacion)")
@@ -56,7 +65,7 @@ class SegmentacionDonacionTest {
     }
 
     // --- Helpers ---
-    private Bien crearBien(Subcategoria sub, Double cantidad, EstadoBien estado, LocalDate vencimiento) {
+    private Bien crearBien(Subcategoria sub, Long cantidad, EstadoBien estado, LocalDate vencimiento) {
         Bien bien = new Bien();
         bien.setSubcategoria(sub);
         bien.setCantidad(cantidad);
@@ -100,7 +109,7 @@ class SegmentacionDonacionTest {
         @Test
         @DisplayName("Un solo bien genera exactamente una donación")
         void unSoloBienGeneraUnaDonacion() {
-            Bien bien = crearBien(subcategoriaFideos, 5.0, null, LocalDate.of(2026, 12, 1));
+            Bien bien = crearBien(subcategoriaFideos, 5L, null, LocalDate.of(2026, 12, 1));
 
             RegistroDonacion registro = crearRegistroConBienes(List.of(bien));
             registro.segmentarDonacion();
@@ -121,8 +130,8 @@ class SegmentacionDonacionTest {
         @DisplayName("Bienes de la misma subcategoría se agrupan en una sola donación")
         void mismaSubcategoriaSeAgrupaEnUnaDonacion() {
             LocalDate vencimiento = LocalDate.of(2026, 12, 1);
-            Bien bien1 = crearBien(subcategoriaFideos, 3.0, null, vencimiento);
-            Bien bien2 = crearBien(subcategoriaFideos, 7.0, null, vencimiento);
+            Bien bien1 = crearBien(subcategoriaFideos, 3L, null, vencimiento);
+            Bien bien2 = crearBien(subcategoriaFideos, 7L, null, vencimiento);
 
             RegistroDonacion registro = crearRegistroConBienes(List.of(bien1, bien2));
             registro.segmentarDonacion();
@@ -136,8 +145,8 @@ class SegmentacionDonacionTest {
         @Test
         @DisplayName("Bienes de distintas subcategorías generan donaciones separadas")
         void distintasSubcategoriasGeneranDonacionesSeparadas() {
-            Bien fideos = crearBien(subcategoriaFideos, 5.0, null, LocalDate.of(2026, 12, 1));
-            Bien leche = crearBien(subcategoriaLeche, 10.0, null, LocalDate.of(2026, 11, 1));
+            Bien fideos = crearBien(subcategoriaFideos, 5L, null, LocalDate.of(2026, 12, 1));
+            Bien leche = crearBien(subcategoriaLeche, 10L, null, LocalDate.of(2026, 11, 1));
 
             RegistroDonacion registro = crearRegistroConBienes(List.of(fideos, leche));
             registro.segmentarDonacion();
@@ -153,22 +162,22 @@ class SegmentacionDonacionTest {
         @Test
         @DisplayName("Perecederos con distinta fecha de vencimiento generan donaciones separadas")
         void perecederosConDistintoVencimientoSeSeparan() {
-            Bien fideos1 = crearBien(subcategoriaFideos, 3.0, null, LocalDate.of(2026, 6, 1));
-            Bien fideos2 = crearBien(subcategoriaFideos, 4.0, null, LocalDate.of(2026, 12, 1));
+            Bien fideos1 = crearBien(subcategoriaFideos, 3L, null, LocalDate.of(2026, 6, 1));
+            Bien fideos2 = crearBien(subcategoriaFideos, 4L, null, LocalDate.of(2026, 12, 1));
 
             RegistroDonacion registro = crearRegistroConBienes(List.of(fideos1, fideos2));
             registro.segmentarDonacion();
 
             assertEquals(2, registro.getDonacionesSegmentadas().size(),
-                "Fideos con diferentes fechas de vencimiento deben generar donaciones separadas");
+                    "Fideos con diferentes fechas de vencimiento deben generar donaciones separadas");
         }
 
         @Test
         @DisplayName("Perecederos con la misma fecha de vencimiento se agrupan")
         void perecederosConMismoVencimientoSeAgrupan() {
             LocalDate mismaFecha = LocalDate.of(2026, 8, 15);
-            Bien fideos1 = crearBien(subcategoriaFideos, 3.0, null, mismaFecha);
-            Bien fideos2 = crearBien(subcategoriaFideos, 7.0, null, mismaFecha);
+            Bien fideos1 = crearBien(subcategoriaFideos, 3L, null, mismaFecha);
+            Bien fideos2 = crearBien(subcategoriaFideos, 7L, null, mismaFecha);
 
             RegistroDonacion registro = crearRegistroConBienes(List.of(fideos1, fideos2));
             registro.segmentarDonacion();
@@ -185,21 +194,21 @@ class SegmentacionDonacionTest {
         @Test
         @DisplayName("Bienes nuevos y usados de la misma subcategoría generan donaciones separadas si la categoría pide estado")
         void nuevosYUsadosSeSeparanSiPideEstado() {
-            Bien sillaNueva = crearBien(subcategoriaSillas, 2.0, EstadoBien.NUEVO, null);
-            Bien sillaUsada = crearBien(subcategoriaSillas, 3.0, EstadoBien.USADO, null);
+            Bien sillaNueva = crearBien(subcategoriaSillas, 2L, EstadoBien.NUEVO, null);
+            Bien sillaUsada = crearBien(subcategoriaSillas, 3L, EstadoBien.USADO, null);
 
             RegistroDonacion registro = crearRegistroConBienes(List.of(sillaNueva, sillaUsada));
             registro.segmentarDonacion();
 
             assertEquals(2, registro.getDonacionesSegmentadas().size(),
-                "Sillas nuevas y usadas deben separarse porque 'Muebles' pideEstado=true");
+                    "Sillas nuevas y usadas deben separarse porque 'Muebles' pideEstado=true");
         }
 
         @Test
         @DisplayName("Bienes con mismo estado de la misma subcategoría se agrupan")
         void mismoEstadoSeAgrupan() {
-            Bien silla1 = crearBien(subcategoriaSillas, 2.0, EstadoBien.NUEVO, null);
-            Bien silla2 = crearBien(subcategoriaSillas, 4.0, EstadoBien.NUEVO, null);
+            Bien silla1 = crearBien(subcategoriaSillas, 2L, EstadoBien.NUEVO, null);
+            Bien silla2 = crearBien(subcategoriaSillas, 4L, EstadoBien.NUEVO, null);
 
             RegistroDonacion registro = crearRegistroConBienes(List.of(silla1, silla2));
             registro.segmentarDonacion();
@@ -220,21 +229,21 @@ class SegmentacionDonacionTest {
             LocalDate venc2 = LocalDate.of(2026, 12, 1);
 
             // 2 fideos con vencimiento junio -> agrupados
-            Bien fideos1 = crearBien(subcategoriaFideos, 3.0, null, venc1);
-            Bien fideos2 = crearBien(subcategoriaFideos, 2.0, null, venc1);
+            Bien fideos1 = crearBien(subcategoriaFideos, 3L, null, venc1);
+            Bien fideos2 = crearBien(subcategoriaFideos, 2L, null, venc1);
 
             // 1 fideos con vencimiento diciembre -> separado
-            Bien fideos3 = crearBien(subcategoriaFideos, 5.0, null, venc2);
+            Bien fideos3 = crearBien(subcategoriaFideos, 5L, null, venc2);
 
             // 1 leche -> separado por subcategoría
-            Bien leche1 = crearBien(subcategoriaLeche, 10.0, null, venc1);
+            Bien leche1 = crearBien(subcategoriaLeche, 10L, null, venc1);
 
             // 2 sillas nuevas -> agrupadas
-            Bien silla1 = crearBien(subcategoriaSillas, 1.0, EstadoBien.NUEVO, null);
-            Bien silla2 = crearBien(subcategoriaSillas, 1.0, EstadoBien.NUEVO, null);
+            Bien silla1 = crearBien(subcategoriaSillas, 1L, EstadoBien.NUEVO, null);
+            Bien silla2 = crearBien(subcategoriaSillas, 1L, EstadoBien.NUEVO, null);
 
             // 1 silla usada -> separada por estado
-            Bien silla3 = crearBien(subcategoriaSillas, 1.0, EstadoBien.USADO, null);
+            Bien silla3 = crearBien(subcategoriaSillas, 1L, EstadoBien.USADO, null);
 
             List<Bien> bienes = List.of(fideos1, fideos2, fideos3, leche1, silla1, silla2, silla3);
             RegistroDonacion registro = crearRegistroConBienes(bienes);
@@ -247,7 +256,7 @@ class SegmentacionDonacionTest {
             // 4. Sillas NUEVO (2 bienes)
             // 5. Sillas USADO (1 bien)
             assertEquals(5, registro.getDonacionesSegmentadas().size(),
-                "Se esperan 5 donaciones segmentadas en este escenario complejo");
+                    "Se esperan 5 donaciones segmentadas en este escenario complejo");
         }
     }
 
@@ -258,22 +267,22 @@ class SegmentacionDonacionTest {
         @Test
         @DisplayName("Todas las donaciones segmentadas arrancan en estado EN_DEPOSITO")
         void donacionesArrancanEnDeposito() {
-            Bien fideos = crearBien(subcategoriaFideos, 5.0, null, LocalDate.of(2026, 12, 1));
-            Bien silla = crearBien(subcategoriaSillas, 2.0, EstadoBien.NUEVO, null);
+            Bien fideos = crearBien(subcategoriaFideos, 5L, null, LocalDate.of(2026, 12, 1));
+            Bien silla = crearBien(subcategoriaSillas, 2L, EstadoBien.NUEVO, null);
 
             RegistroDonacion registro = crearRegistroConBienes(List.of(fideos, silla));
             registro.segmentarDonacion();
 
             for (Donacion donacion : registro.getDonacionesSegmentadas()) {
                 assertEquals(TipoEstadoDonacion.EN_DEPOSITO, donacion.estadoActual(),
-                    "Cada donación segmentada debe arrancar EN_DEPOSITO");
+                        "Cada donación segmentada debe arrancar EN_DEPOSITO");
             }
         }
 
         @Test
         @DisplayName("Las donaciones segmentadas tienen historial de estados con exactamente 1 entrada")
         void donacionesTienenUnEstadoEnHistorial() {
-            Bien bien = crearBien(subcategoriaFideos, 5.0, null, LocalDate.of(2026, 12, 1));
+            Bien bien = crearBien(subcategoriaFideos, 5L, null, LocalDate.of(2026, 12, 1));
 
             RegistroDonacion registro = crearRegistroConBienes(List.of(bien));
             registro.segmentarDonacion();
@@ -285,7 +294,7 @@ class SegmentacionDonacionTest {
         @Test
         @DisplayName("La fecha de la donación segmentada es la fecha del registro")
         void fechaDonacionEsLaDelRegistro() {
-            Bien bien = crearBien(subcategoriaFideos, 5.0, null, LocalDate.of(2026, 12, 1));
+            Bien bien = crearBien(subcategoriaFideos, 5L, null, LocalDate.of(2026, 12, 1));
 
             RegistroDonacion registro = crearRegistroConBienes(List.of(bien));
             registro.segmentarDonacion();
